twitch_chat_alert/requirements.txt
python-dotenv==1.0.0
playsound==1.2.2
